requirements.txt
slackclient
apache-airflow[gcp_api,postgres]==1.10.6
pytest
pytest-mock